# Conflicts:
#	SolrNet/Impl/FieldParsers/AggregateFieldParser.cs

+++ SolrNet/Impl/AbstractSolrQueryResults.cs
@@ -6,19 +6,19 @@ namespace SolrNet.Impl {
     public abstract class AbstractSolrQueryResults<T> : List<T> {
         /// <summary>
         /// CursorMark token returned for deep pagination.
-        /// Only present if explicitly requested through <see cref="SolrNet.Commands.Parameters.CommonQueryOptions.StartOrCursor"/>
+        /// Only present if explicitly requested through <see cref="StartOrCursor"/>
         /// </summary>
         public StartOrCursor.Cursor NextCursorMark { get; set; }
 
         /// <summary>
         /// Total documents found
         /// </summary>
-        public int NumFound { get; set; }
+        public long NumFound { get; set; }
 
         /// <summary>
         /// Start of the results
         /// </summary>
-        public int Start { get; set; }
+        public long Start { get; set; }
 
         /// <summary>
         /// Max score in these results
