+++ SolrNet/Impl/ResponseParsers/ResultsResponseParser.cs
@@ -60,13 +60,14 @@ namespace SolrNet.Impl.ResponseParsers {
             return new StartOrCursor.Cursor(nextCursorMarkElement.Value);
         }
 
+        /// <inheritdoc/>
         public void Parse(XDocument xml, AbstractSolrQueryResults<T> results) {
             var resultNode = GetMainResultNode(xml) ?? GetGroupResultNode(xml);
             if (resultNode == null)
                 return;
 
-            results.NumFound = Convert.ToInt32(resultNode.Attribute("numFound").Value);
-            results.Start = Convert.ToInt32(resultNode.Attribute("start").Value);
+            results.NumFound = Convert.ToInt64(resultNode.Attribute("numFound").Value);
+            results.Start = Convert.ToInt64(resultNode.Attribute("start").Value);
             var maxScore = resultNode.Attribute("maxScore");
             if (maxScore != null)
                 results.MaxScore = double.Parse(maxScore.Value, CultureInfo.InvariantCulture.NumberFormat);
